Dup options in iniatilizer
Typically, "best practice" is to dup an option coming in
so it doesn't effect anyone.

But to be honest, I'm not sure what I feel about this. Seems
wasteful. And most of the time the options didn't change so it
is just paranoia and wasteful.

In this case, our intent is to modify the options hash. So it actually
feels pertinent here. Most of the time the options will be set from
ruby code. But I have no idea how this code will be used.